backend/requirements.txt
flask
langchain
tidb-vector
langchain-core
langchain-community
pymysql
langchain-google-genai
pandas
python-dotenv
markdownify
